Extract models to their own separate classes
